amethyx-main/requirements.txt
hikari
hikari-lightbulb
hikari-miru
aiohttp
lightbulb-ext-tasks
pillow
python-dotenv
motor
pymongo
